adjust data in recipe page
