+++ client/src/pages/Recipe.tsx
@@ -4,9 +4,10 @@ import { QueryClient, useQuery } from '@tanstack/react-query';
 import { waveform } from 'ldrs';
 import { FaPrint, FaSave, FaShareAlt, FaStar } from 'react-icons/fa';
 import { Params, useParams } from 'react-router-dom';
-import ImageSlide, { Dish } from '../components/meals/ImageSlide';
+import ImageSlide from '../components/meals/ImageSlide';
 import customFetch from '../utils/customFetch';
 import { Tag, TagLabel } from '@chakra-ui/react';
+import { BackendData } from '../components/meals/ImageSlide';
 
 waveform.register();
 
@@ -80,217 +81,198 @@ export const loader =
 //     console.log(recipeData);
 //     console.log(status);
 
-
-interface MealProps {
-    individualMeal: Dish[];
-    title: string;
-    overview: string;
-    image: string;
-    source: string;
-    totalTime: string;
-    servings: string;
+interface IndividualMealProps {
+    recipeData: BackendData;
     calories: string;
     averageStar: string;
     numReviews: string;
-    tags: string[];
-};
+    }
 
-function IndividualMeal({ 
-    individualMeal, 
-    title, 
-    overview,
-    image, 
-    source,
-    totalTime, 
-    servings, 
-    calories, 
-    averageStar, 
-    numReviews,
-    tags }: MealProps) {
-    
+const IndividualMeal: React.FC<IndividualMealProps> = ({ recipeData, calories, averageStar, numReviews }) => {
+    const totalTime = recipeData.analyzeInstruction.reduce((acc, instruction) => {
+        return (
+        acc +
+        instruction.steps.reduce((stepAcc, step) => {
+            return stepAcc + (step.length?.number || 0);
+        }, 0)
+        );
+    }, 0);
+
     return (
         <Stack alignItems="center" justifyContent="center">
-            <Grid templateRows="repeat(5)" templateColumns="repeat(4, 2fr)" mt="6" width="100%">
-                <GridItem rowSpan={5} colSpan={2} ml={4} objectFit="cover">
-                    <Box
-                        width="100%"
-                        height="100%"
-                        display="flex"
-                        justifyContent="center"
-                        alignItems="center"
-                        borderRadius="md"
-                        mb="3"
-                        objectFit="cover"
-                        position="relative"
-                    >
-                        <Image src={image} alt={title} objectFit="cover" borderRadius="md"/>
-                        <Box
-                        position="absolute"
-                        bottom="0"
-                        left="0"
-                        right="0"
-                        p={4}
-                        display="flex"
-                        flexWrap="wrap"
-                        justifyContent="left"
-                        alignItems="center"
-                        >
-                        {tags.map((tag, index) => (
-                            <Tag
-                            key={index}
-                            bg="rgba(126, 126, 126, 0.4)" 
-                            color="white"
-                            borderRadius="full"
-                            fontSize="20"
-                            m={1}
-                            >
-                                <TagLabel px={1} py={1}>
-                                    {tag}
-                                </TagLabel>
-                            </Tag>
-                        ))}
-                        </Box>
-                    </Box>
-                </GridItem>
-                <GridItem
-                    rowSpan={4}
-                    colSpan={2}
-                    fontSize={34}
-                    display="flex"
-                    flexDirection="column"
-                    justifyContent="center"
-                    alignItems="center"
+        <Grid templateRows="repeat(5)" templateColumns="repeat(4, 2fr)" mt="6" width="100%">
+            <GridItem rowSpan={5} colSpan={2} ml={4} objectFit="cover">
+            <Box
+                width="100%"
+                height="100%"
+                display="flex"
+                justifyContent="center"
+                alignItems="center"
+                borderRadius="md"
+                mb="3"
+                objectFit="cover"
+                position="relative"
+            >
+                <Image src={recipeData.imageUrl} alt={recipeData.title} objectFit="cover" borderRadius="md" />
+                <Box
+                position="absolute"
+                bottom="0"
+                left="0"
+                right="0"
+                p={4}
+                display="flex"
+                flexWrap="wrap"
+                justifyContent="left"
+                alignItems="center"
                 >
-                    <Box ml="10" mr="10" textAlign="center" alignItems="center" display="flex" flexDirection="column">
-                    <Box
-                        width="80px"
-                        height="36px"
-                        bg="black"
-                        textColor="white"
-                        fontSize="14"
-                        display="flex"
-                        flexDirection="column"
-                        justifyContent="center"
-                        alignItems="center"
-                        mb={3}
-                        textAlign="center"
-                    >
-                        RECIPES
-                    </Box>
-                    <Heading
-                        mb="3"
-                        mt="4"
-                        size="lg"
-                        fontSize="50"
-                        fontFamily="'Playfair Display', serif"
-                        fontWeight="bold"
-                        textAlign="center"
-                    >
-                        {title}
-                    </Heading>
-                    <Text justifyContent="center" alignItems="center" fontSize={26}>
-                        {overview}
-                    </Text>
-                    <HStack mt="2" justifyContent="center" alignItems="center" fontSize="14">
-                        <Text fontSize={18} ml={2}>
-                        {averageStar}
-                        </Text>
-                        {Array(5)
-                        .fill('')
-                        .map((_, i) => (
-                            <FaStar key={i} color="black" />
-                        ))}
-                        <Text fontSize={18} ml={2}>
-                        ({numReviews})
-                        </Text>
-                    </HStack>
-                    <HStack justifyContent="flex-end" width="100%" mt={8}>
-                        <ButtonGroup spacing={2} mb={10}>
-                        <Button rightIcon={<FaSave />} bg="base.50">
-                            Save
-                        </Button>
-                        <Button rightIcon={<FaStar />} bg="base.50">
-                            Rate
-                        </Button>
-                        <Button rightIcon={<FaPrint />} bg="base.50">
-                            Print
-                        </Button>
-                        <Button rightIcon={<FaShareAlt />} bg="base.50">
-                            Share
-                        </Button>
-                        </ButtonGroup>
-                    </HStack>
-                    <Text fontSize="14" textAlign="center" textColor="base.600">
-                        * Source: {source}
-                    </Text>
-                    </Box>
-                </GridItem>
-                <GridItem colSpan={2}>
-                    <Box h="100%" display="flex" flexDirection="column" justifyContent="flex-end" alignItems="center">
-                    <HStack justifyContent="center" alignItems="center" width="100%" spacing={0}>
-                        <Box
-                        bg="base.50"
-                        p={4}
-                        position="relative"
-                        width="25%"
-                        _after={{
-                            content: '""',
-                            position: 'absolute',
-                            right: 0,
-                            top: '20%',
-                            bottom: '20%',
-                            width: '1px',
-                            bg: 'base.200',
-                        }}
-                        >
-                        <Text fontSize="18">
-                            <Box as="span" fontSize="20" fontWeight="bold">
-                            Total time:
-                            </Box>
-                            <Box>{totalTime}</Box>
-                        </Text>
-                        </Box>
-                        <Box
-                        bg="base.50"
-                        p={4}
-                        position="relative"
-                        width="25%"
-                        _after={{
-                            content: '""',
-                            position: 'absolute',
-                            right: 0,
-                            top: '20%',
-                            bottom: '20%',
-                            width: '1px',
-                            bg: 'base.200',
-                        }}
-                        >
-                        <Text fontSize="18">
-                            <Box as="span" fontSize="20" fontWeight="bold">
-                            Servings:
-                            </Box>
-                            <Box>{servings}</Box>
-                        </Text>
-                        </Box>
-                        <Box bg="base.50" p={4} width="40%">
-                        <Text fontSize="18">
-                            <Box as="span" fontSize="20" fontWeight="bold">
-                            Calories per serving:
-                            </Box>{' '}
-                            <Box>{calories}</Box>
-                        </Text>
-                        </Box>
-                    </HStack>
-                    </Box>
-                </GridItem>
-            </Grid>
-            <Box width="100%" mx="auto" px={4}>
-                <ImageSlide dishes={individualMeal} />
+                {recipeData.tags.map((tag, index) => (
+                    <Tag key={index} bg="rgba(126, 126, 126, 0.4)" color="white" borderRadius="full" fontSize="20" m={1}>
+                    <TagLabel px={1} py={1}>
+                        {tag}
+                    </TagLabel>
+                    </Tag>
+                ))}
+                </Box>
             </Box>
+            </GridItem>
+            <GridItem
+            rowSpan={4}
+            colSpan={2}
+            fontSize={34}
+            display="flex"
+            flexDirection="column"
+            justifyContent="center"
+            alignItems="center"
+            >
+            <Box ml="10" mr="10" textAlign="center" alignItems="center" display="flex" flexDirection="column">
+                <Box
+                width="80px"
+                height="36px"
+                bg="black"
+                textColor="white"
+                fontSize="14"
+                display="flex"
+                flexDirection="column"
+                justifyContent="center"
+                alignItems="center"
+                mb={3}
+                textAlign="center"
+                >
+                RECIPES
+                </Box>
+                <Heading
+                mb="3"
+                mt="4"
+                size="lg"
+                fontSize="50"
+                fontFamily="'Playfair Display', serif"
+                fontWeight="bold"
+                textAlign="center"
+                >
+                {recipeData.title}
+                </Heading>
+                <Text justifyContent="center" alignItems="center" fontSize={26}>
+                {recipeData.description}
+                </Text>
+                <HStack mt="2" justifyContent="center" alignItems="center" fontSize="14">
+                <Text fontSize={18} ml={2}>
+                    {averageStar}
+                </Text>
+                {Array(5)
+                    .fill('')
+                    .map((_, i) => (
+                    <FaStar key={i} color="black" />
+                    ))}
+                <Text fontSize={18} ml={2}>
+                    ({numReviews})
+                </Text>
+                </HStack>
+                <HStack justifyContent="flex-end" width="100%" mt={8}>
+                <ButtonGroup spacing={2} mb={10}>
+                    <Button rightIcon={<FaSave />} bg="base.50">
+                    Save
+                    </Button>
+                    <Button rightIcon={<FaStar />} bg="base.50">
+                    Rate
+                    </Button>
+                    <Button rightIcon={<FaPrint />} bg="base.50">
+                    Print
+                    </Button>
+                    <Button rightIcon={<FaShareAlt />} bg="base.50">
+                    Share
+                    </Button>
+                </ButtonGroup>
+                </HStack>
+                <Text fontSize="14" textAlign="center" textColor="base.600">
+                * Source: {recipeData.source}
+                </Text>
+            </Box>
+            </GridItem>
+            <GridItem colSpan={2}>
+            <Box h="100%" display="flex" flexDirection="column" justifyContent="flex-end" alignItems="center">
+                <HStack justifyContent="center" alignItems="center" width="100%" spacing={0}>
+                <Box
+                    bg="base.50"
+                    p={4}
+                    position="relative"
+                    width="25%"
+                    _after={{
+                    content: '""',
+                    position: 'absolute',
+                    right: 0,
+                    top: '20%',
+                    bottom: '20%',
+                    width: '1px',
+                    bg: 'base.200',
+                    }}
+                >
+                    <Text fontSize="18">
+                    <Box as="span" fontSize="20" fontWeight="bold">
+                        Total time:
+                    </Box>
+                    <Box>{totalTime} minutes</Box>
+                    </Text>
+                </Box>
+                <Box
+                    bg="base.50"
+                    p={4}
+                    position="relative"
+                    width="25%"
+                    _after={{
+                    content: '""',
+                    position: 'absolute',
+                    right: 0,
+                    top: '20%',
+                    bottom: '20%',
+                    width: '1px',
+                    bg: 'base.200',
+                    }}
+                >
+                    <Text fontSize="18">
+                    <Box as="span" fontSize="20" fontWeight="bold">
+                        Servings:
+                    </Box>
+                    <Box>{recipeData.servings}</Box>
+                    </Text>
+                </Box>
+                <Box bg="base.50" p={4} width="40%">
+                    <Text fontSize="18">
+                    <Box as="span" fontSize="20" fontWeight="bold">
+                        Calories per serving:
+                    </Box>{' '}
+                    <Box>{calories}</Box>
+                    </Text>
+                </Box>
+                </HStack>
+            </Box>
+            </GridItem>
+        </Grid>
+        <Box width="100%" mx="auto" px={4}>
+            <ImageSlide backendData={recipeData} />
+        </Box>
         </Stack>
     );
-}
+};
 
 // export default Recipe;
-
 export default IndividualMeal;
